requirements.txt
bcrypt==3.2.0
mysql-connector-python==8.0.28
python-decouple==3.6
pyjwt==2.3.0
pytest==7.1.2
fastapi[all]==0.75.2